feat: add player class
